Fix stream test_recorder.py flakiness
Fix stream tests to no longer be flaky, and bring tests up to day with
changes that have happened since they were initally disabled.

Breakdown of changes:
- Fix test to pass a stream instead of a BytesIO given changes in the
  recorder that assume a stream path is passed in
- Share test fixture with test_hls.py for synchronization with worker
- Fix race conditions where the stream worker thread may exit before the
  test can consume streams
- Fix teardown thread leaks by running all event loop callbacks that may
  invoke cleanup under patch
